+++ server/models/index.js
@@ -1,0 +1,7 @@
+const User = require('./User');
+const Location = require('./Location');
+const Instrument = require('./Instrument');
+const Filter = require('./Filter');
+const Session = require('./Session');
+
+module.exports = { User, Location, Instrument, Filter, Session };